Renamed controlling field
A field name of the object it is referencing is okay if it's a different object. But having the object name as a field on the object was confusing things. Changed all references on AP and APT to Controller__c and TController__c.
Also updated global value set API names in case we have an issue with exporting/importing/anonymous apex.